+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 ### Improved
+- Update emotion plugin to support 11 version ([#1558](https://github.com/react-static/react-static/pull/1558))
 
 ### Bugfix
 Fix publicPath is used for webpack output publicPath instead of assetsPath ([#1569](https://github.com/react-static/react-static/pull/1569))